Provided extensibility for n puzzler solver
